+++ packages/client/src/remote-invoke.ts
@@ -4,14 +4,13 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import type { RemoteFunction, Serialization } from "@quatico/magellan-shared";
-import { deserialize, packInput } from "@quatico/magellan-shared";
+import { deserialize, packInput, RemoteFunction, Serialization } from "@quatico/magellan-shared";
 import type { Context } from "./transport";
 import { transportRequest } from "./transport";
 
 export const remoteInvoke = async <O>(
     func: RemoteFunction,
-    ctx: Context = { headers: new Headers() },
+    ctx: Context = { headers: {} },
     serialization: Serialization = { serialize: packInput, deserialize }
 ): Promise<O> => {
     return transportRequest(func, ctx, serialization);