+++ packages/client/src/transport/transport-request.spec.ts
@@ -5,20 +5,16 @@
  * ---------------------------------------------------------------------------------------------
  */
 
-import { packInput, serialize } from "@quatico/magellan-shared";
+import { packInput } from "@quatico/magellan-shared";
 import { initProjectConfiguration } from "./configuration-repository";
 import { formdataFetch } from "./formdata-fetch";
 import { transportRequest } from "./transport-request";
 
 let formAppend: jest.Mock;
 let formSet: jest.Mock;
-let headerAppend: jest.Mock;
-let headerSet: jest.Mock;
 beforeAll(() => {
     formAppend = jest.fn();
     formSet = jest.fn();
-    headerAppend = jest.fn();
-    headerSet = jest.fn();
 
     function formDataMock() {
         // @ts-ignore
@@ -28,15 +24,6 @@ beforeAll(() => {
     }
     // @ts-ignore
     global.FormData = formDataMock;
-
-    function headerMock() {
-        // @ts-ignore
-        this.append = headerAppend;
-        // @ts-ignore
-        this.set = headerSet;
-    }
-    // @ts-ignore
-    global.Headers = headerMock;
 });
 
 describe("transportRequest", () => {
@@ -46,7 +33,7 @@ describe("transportRequest", () => {
     });
 
     it("calls fetch passing serialized input parameters with simple object", async () => {
-        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => serialize(target) }));
+        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => packInput(target) }));
         const target = { data: "whatever" };
 
         await transportRequest({ name: "whatever", data: target });
@@ -54,7 +41,7 @@ describe("transportRequest", () => {
         expect(fetch).toHaveBeenCalledTimes(1);
         expect(fetch).toHaveBeenCalledWith(`/api`, {
             body: expect.any(FormData),
-            headers: expect.any(Headers),
+            headers: expect.anything(),
             method: "POST",
         });
 
@@ -66,14 +53,14 @@ describe("transportRequest", () => {
 
     it("calls fetch using no input parameter w/o data property", async () => {
         const target = {};
-        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => serialize(target) }));
+        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => packInput(target) }));
 
         await transportRequest({ name: "whatever" });
 
         expect(global.fetch).toHaveBeenCalledTimes(1);
         expect(global.fetch).toHaveBeenCalledWith(`/api`, {
             body: expect.any(FormData),
-            headers: expect.any(Headers),
+            headers: expect.anything(),
             method: "POST",
         });
 
@@ -85,11 +72,11 @@ describe("transportRequest", () => {
 
     it("calls deserialize passing result from fetch", async () => {
         global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => Buffer.from("expected") }));
-        const target = jest.fn();
+        const target = jest.fn().mockImplementation(value => value);
 
         await transportRequest(
             { name: "whatever", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: jest.fn(),
                 deserialize: target,
@@ -100,15 +87,15 @@ describe("transportRequest", () => {
     });
 
     it("calls serialize passing result from fetch", async () => {
-        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => serialize("expected") }));
+        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => packInput("expected") }));
         const target = jest.fn();
 
         await transportRequest(
             { name: "whatever", data: "expected" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: target,
-                deserialize: jest.fn(),
+                deserialize: jest.fn().mockImplementation(value => value),
             }
         );
 
@@ -116,7 +103,7 @@ describe("transportRequest", () => {
     });
 
     it("calls fetch passing function name from input", async () => {
-        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => serialize("whatever") }));
+        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => packInput("whatever") }));
         initProjectConfiguration({
             namespaces: { default: { endpoint: "http://expected-host:3000/api" } },
             transports: { default: formdataFetch },
@@ -124,7 +111,7 @@ describe("transportRequest", () => {
 
         await transportRequest(
             { name: "expected", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: val => JSON.stringify(val),
                 deserialize: val => JSON.parse(val),
@@ -134,7 +121,7 @@ describe("transportRequest", () => {
         expect(fetch).toHaveBeenCalledTimes(1);
         expect(fetch).toHaveBeenCalledWith(`http://expected-host:3000/api`, {
             body: expect.any(FormData),
-            headers: expect.any(Headers),
+            headers: expect.anything(),
             method: "POST",
         });
 
@@ -145,7 +132,7 @@ describe("transportRequest", () => {
     });
 
     it("calls fetch w/ baseUrl, servicePath", async () => {
-        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => serialize("whatever") }));
+        global.fetch = jest.fn().mockReturnValue(Promise.resolve({ text: () => packInput("whatever") }));
         initProjectConfiguration({
             namespaces: { default: { endpoint: "http://expected-host:3000/expected/path/function" } },
             transports: { default: formdataFetch },
@@ -153,7 +140,7 @@ describe("transportRequest", () => {
 
         await transportRequest(
             { name: "target", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: val => JSON.stringify(val),
                 deserialize: val => JSON.parse(val),
@@ -163,7 +150,7 @@ describe("transportRequest", () => {
         expect(fetch).toHaveBeenCalledTimes(1);
         expect(fetch).toHaveBeenCalledWith(`http://expected-host:3000/expected/path/function`, {
             body: expect.any(FormData),
-            headers: expect.any(Headers),
+            headers: expect.anything(),
             method: "POST",
         });
 
@@ -184,7 +171,7 @@ describe("transportRequest", () => {
 
         const actual = transportRequest(
             { name: "foobar", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: () => {
                     throw new Error("expected");