+++ packages/server/src/routes/function-route.spec.ts
@@ -39,11 +39,11 @@ describe("createStaticRoute", () => {
             .field("data", JSON.stringify({ input: "whatever" }));
 
         expect(res.header["content-type"]).toBe("application/json; charset=utf-8");
-        expect(unpackPayload(res.body)).toBe("Hello World!");
+        expect(unpackPayload(res.body).data).toBe("Hello World!");
         expect(res.statusCode).toBe(200);
     });
 
-    it("responds to /api with registered function that throws error", async () => {
+    it("responds to /api with registered function that throws error and includes error field in non-production", async () => {
         app.use(
             "/api",
             createFunctionRoute(
@@ -63,11 +63,38 @@ describe("createStaticRoute", () => {
             .field("data", JSON.stringify({ input: "whatever" }));
 
         expect(res.header["content-type"]).toBe("application/json; charset=utf-8");
-        const { error, message } = unpackPayload(res.body) as { error: string; message: string };
-        expect(message).toBe('Function request to "expected" failed.');
-        expect(error.toString()).toBe("expected error message");
-        expect(res.error.toString()).toBe("Error: cannot POST /api (500)");
-        expect(res.statusCode).toBe(500);
+        const payload = unpackPayload(res.body);
+        expect(payload.error).toEqual({
+            message: 'Function request to "expected" failed.',
+            error: "expected error message",
+        });
+        expect(res.statusCode).toBe(200);
+    });
+
+    it("responds to /api with registered function that throws error and omits error field in production", async () => {
+        app.use(
+            "/api",
+            createFunctionRoute(
+                new Sdk().registerFunction(
+                    "expected",
+                    jest.fn().mockImplementation(() => {
+                        throw new Error("expected error message");
+                    })
+                )
+            )
+        );
+        process.env.NODE_ENV = "production";
+
+        const res = await request(app)
+            .post("/api")
+            .set("Accept", "application/json")
+            .field("name", "expected")
+            .field("data", JSON.stringify({ input: "whatever" }));
+
+        expect(res.header["content-type"]).toBe("application/json; charset=utf-8");
+        const payload = unpackPayload(res.body);
+        expect(payload.error).toEqual({ message: 'Function request to "expected" failed.' });
+        expect(res.statusCode).toBe(200);
     });
 
     // TODO: This assumption no longer holds until a full function registration with name and namespace becomes available across frontend, node and java