+++ packages/server/src/transport/transport-request.spec.ts
@@ -5,8 +5,7 @@
  * ---------------------------------------------------------------------------------------------
  */
 
-import { packInput, serialize } from "@quatico/magellan-shared";
-import { Headers } from "node-fetch";
+import { serialize } from "@quatico/magellan-shared";
 import { addNamespace, addTransport, initProjectConfiguration } from "../configuration";
 import { initDependencyContext } from "../services";
 import { transportRequest } from "./transport-request";
@@ -37,8 +36,8 @@ describe("transportRequest", () => {
 
         expect(transportHandler).toHaveBeenCalledTimes(1);
         expect(transportHandler).toHaveBeenCalledWith(
-            { endpoint: "/api", payload: packInput(target), name: "whatever", namespace: "test" },
-            { headers: new Headers() }
+            { endpoint: "/api", payload: serialize(target), name: "whatever", namespace: "test" },
+            { headers: {} }
         );
     });
 
@@ -50,37 +49,37 @@ describe("transportRequest", () => {
 
         expect(transportHandler).toHaveBeenCalledTimes(1);
         expect(transportHandler).toHaveBeenCalledWith(
-            { endpoint: "/api", payload: packInput(target), name: "whatever", namespace: "test" },
-            { headers: new Headers() }
+            { endpoint: "/api", payload: serialize(target), name: "whatever", namespace: "test" },
+            { headers: {} }
         );
     });
 
     it("calls deserialize passing result from fetch", async () => {
-        transportHandler.mockReturnValue(Buffer.from("expected"));
-        const target = jest.fn();
+        transportHandler.mockReturnValue({ data: Buffer.from("expected") });
+        const target = jest.fn().mockImplementation(value => value);
 
         await transportRequest(
             { name: "whatever", data: "whatever", namespace: "test" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: jest.fn(),
                 deserialize: target,
             }
         );
 
-        expect(target).toHaveBeenCalledWith(Buffer.from("expected"));
+        expect(target).toHaveBeenCalledWith({ data: Buffer.from("expected") });
     });
 
     it("calls serialize passing result from fetch", async () => {
-        transportHandler.mockReturnValue(serialize("expected"));
+        transportHandler.mockReturnValue({ data: serialize("expected") });
         const target = jest.fn();
 
         await transportRequest(
             { name: "whatever", data: "expected", namespace: "test" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: target,
-                deserialize: jest.fn(),
+                deserialize: jest.fn().mockImplementation(value => value),
             }
         );
 
@@ -96,7 +95,7 @@ describe("transportRequest", () => {
 
         await transportRequest(
             { name: "expected", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: (val: unknown) => JSON.stringify(val),
                 deserialize: (val: string) => JSON.parse(val),
@@ -106,7 +105,7 @@ describe("transportRequest", () => {
         expect(transportHandler).toHaveBeenCalledTimes(1);
         expect(transportHandler).toHaveBeenCalledWith(
             { endpoint: "http://expected-host:3000/api", payload: JSON.stringify("whatever"), name: "expected", namespace: "default" },
-            { headers: new Headers() }
+            { headers: {} }
         );
     });
 
@@ -119,7 +118,7 @@ describe("transportRequest", () => {
 
         await transportRequest(
             { name: "target", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: (val: unknown) => JSON.stringify(val),
                 deserialize: (val: string) => JSON.parse(val),
@@ -134,7 +133,7 @@ describe("transportRequest", () => {
                 name: "target",
                 namespace: "default",
             },
-            { headers: new Headers() }
+            { headers: {} }
         );
     });
 
@@ -149,7 +148,7 @@ describe("transportRequest", () => {
 
         const actual = transportRequest(
             { name: "foobar", data: "whatever" },
-            { headers: new Headers() },
+            { headers: {} },
             {
                 serialize: () => {
                     throw new Error("expected");
@@ -166,6 +165,6 @@ describe("transportRequest", () => {
 
         const actual = transportRequest({ name: "whatever", data: "whatever", namespace: "test" });
 
-        await expect(actual).rejects.toThrow('Cannot deserialize response from remote function: "whatever".');
+        await expect(actual).rejects.toThrow('Cannot deserialize response from invoke function: "whatever".');
     });
 });