+++ packages/server/src/transport/transport-request.ts
@@ -4,15 +4,14 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
-import { deserialize, packInput, RemoteFunction, Serialization } from "@quatico/magellan-shared";
-import { Headers } from "node-fetch";
+import { deserialize, RemoteFunction, Serialization, serialize } from "@quatico/magellan-shared";
 import type { Context } from "../api";
 import { resolveNamespace } from "../configuration";
 
 export const transportRequest = async <O>(
     func: RemoteFunction,
-    ctx: Context = { headers: new Headers() },
-    serialization: Serialization = { serialize: packInput, deserialize }
+    ctx: Context = { headers: {} },
+    serialization: Serialization = { serialize, deserialize }
 ): Promise<O> => {
     const { name, data = {}, namespace = "default" } = func;
 
@@ -31,8 +30,13 @@ export const transportRequest = async <O>(
 
     try {
         const result = await transport({ name, payload, namespace, endpoint }, ctx);
-        return serialization.deserialize(result);
+        const { data, error } = serialization.deserialize(result);
+        if (error?.error) {
+            // eslint-disable-next-line no-console
+            console.error(`Function request to "${name}" failed with error ${error.error}`);
+        }
+        return error ? Promise.reject(error.message) : (data as O);
     } catch (err) {
-        throw new Error(`Cannot deserialize response from remote function: "${name}". Reason: "${err}".`);
+        throw new Error(`Cannot deserialize response from invoke function: "${name}". Reason: "${err}".`);
     }
 };