+++ packages/shared/src/transport/Serialization.ts
@@ -4,7 +4,8 @@
  *   Licensed under the MIT License. See LICENSE in the project root for license information.
  * ---------------------------------------------------------------------------------------------
  */
+import { ResponsePayload } from "./ResponsePayload";
 export interface Serialization {
-    deserialize: <O>(json: string) => O;
+    deserialize: <O>(json: string) => ResponsePayload<O>;
     serialize: <I>(input: I) => string;
 }