+++ packages/test/src/steps/magellan-cli.steps.ts
@@ -1,14 +1,18 @@
+import { After, Before, Given, Then, When } from "@cucumber/cucumber";
+import assert from "assert";
 import { copyFileSync, existsSync, mkdirSync, rmSync, writeFileSync } from "fs";
 import { basename, extname, isAbsolute, join, resolve } from "path";
 import { Cli } from "../cli";
-import { After, Before, Given, Then, When } from "@cucumber/cucumber";
-import assert from "assert";
 
-type CliOperation = "compile";
+type CliOperation = "compile" | "serve";
 let previousPath = ".";
 const targetDirectory = resolve("output", "project");
 const dataPath = resolve("test", "__data__");
+let remoteInvokeResult: unknown;
+let remoteInvokeError: unknown;
+let remoteInvokeConsoleError: string | undefined;
 let projectDirectory: string = targetDirectory;
+let cli: Cli;
 
 Given(/^valid TypeScript project directory was created$/, () => {
     mkdirSync(targetDirectory, { recursive: true });
@@ -39,35 +43,121 @@ Given(/^valid index module file "(.*)" was created$/, fileName => {
     copyFileSync(resolve(dataPath, "index-modules", fileName), resolve(targetDirectory, previousPath, "index.ts"));
 });
 
-When(/^CLI command "(.*)" is called without arguments$/, async (cliOperation: CliOperation) => {
+Given("node environment is {string}", (env: string) => {
+    process.env.NODE_ENV = env;
+});
+
+When(/^CLI command "(.*)" is called without arguments$/, { timeout: 120 * 1000 }, async (cliOperation: CliOperation) => {
     process.chdir(targetDirectory);
-    // eslint-disable-next-line no-console
-    console.warn(`Current working directory: ${process.cwd()}`);
-    return cliOperation === "compile" ? new Cli().executeCompile({}) : Promise.reject(`Unsupported cli operation ${cliOperation}`);
+    cli = cli || new Cli();
+    switch (cliOperation) {
+        case "compile":
+            return cli.executeCompile({});
+        case "serve": {
+            const { config } = await cli.executeServe({ command: { args: "-s ./lib/server ./lib/client " } });
+            // Because cucumber-js runs in the node environment, we need to tell the 'frontend' on what host it is running.
+            // Also, because it is node that executes it, we must use the servers default transport on the frontend to have it use node-fetch!
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (global as any).__qsMagellanConfig__ = config;
+            break;
+        }
+        default:
+            return Promise.reject(`Unsupported cli operation ${cliOperation}`);
+    }
 });
 
 When(/^CLI command "(.*)" is called with arguments "(.*)"$/, async (cliOperation: CliOperation, args: string) => {
     process.chdir(targetDirectory);
-    return cliOperation === "compile"
-        ? new Cli().executeCompile({
-              command: {
-                  args: args !== "" ? args.replaceAll(/<projectDir>/g, targetDirectory) : undefined,
-                  cwd: resolve(targetDirectory),
-              },
-          })
-        : Promise.reject(`Unsupported cli operation ${cliOperation}`);
+
+    cli = cli || new Cli();
+    switch (cliOperation) {
+        case "compile":
+            return cli.executeCompile({
+                command: {
+                    args: args !== "" ? args.replaceAll(/<projectDir>/g, targetDirectory) : undefined,
+                    cwd: resolve(targetDirectory),
+                },
+            });
+        case "serve": {
+            const { config } = await cli.executeServe({
+                command: {
+                    // args must include at least "-s ./lib/server ./lib/client "
+                    args: args !== "" ? args.replaceAll(/<projectDir>/g, targetDirectory) : undefined,
+                    cwd: resolve(targetDirectory),
+                },
+            });
+            // Because cucumber-js runs in the node environment, we need to tell the 'frontend' on what host it is running.
+            // Also, because it is node that executes it, we must use the servers default transport on the frontend to have it use node-fetch!
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (global as any).__qsMagellanConfig__ = config;
+            break;
+        }
+        default:
+            return Promise.reject(`Unsupported cli operation ${cliOperation}`);
+    }
+});
+
+When("the function {string} is invoked", { timeout: 60 * 1000 }, async function (functionName: string) {
+    try {
+        const module = await import(join(process.cwd(), "lib", "client", functionName));
+        // eslint-disable-next-line no-console
+        const errorConsole = console.error;
+        // eslint-disable-next-line no-console
+        console.error = error => (remoteInvokeConsoleError = error);
+        remoteInvokeResult = await module[functionName]();
+        // eslint-disable-next-line no-console
+        console.error = errorConsole;
+    } catch (e) {
+        remoteInvokeError = e;
+    }
+});
+
+When("the function {string} is invoked with {string}", { timeout: 60 * 1000 }, async function (functionName: string, data: string) {
+    try {
+        const module = await import(join(process.cwd(), "lib", "client", functionName));
+        // eslint-disable-next-line no-console
+        const errorConsole = console.error;
+        // eslint-disable-next-line no-console
+        console.error = error => (remoteInvokeConsoleError = error);
+        remoteInvokeResult = await module[functionName](JSON.parse(data));
+        // eslint-disable-next-line no-console
+        console.error = errorConsole;
+    } catch (e) {
+        remoteInvokeError = e;
+    }
 });
 
 Then(/^directory "(.*)" contains file "(.*)"$/, (directory, fileName) => {
     assert.equal(existsSync(resolve(targetDirectory, directory, fileName)), true, `Directory ${directory} should contain file ${fileName}`);
 });
 
+Then("the promise is resolved with {string}.", (result: string) => {
+    assert.equal(result, JSON.stringify(remoteInvokeResult));
+});
+
+Then("the promise is rejected with message {string}.", (message: string) => {
+    assert.equal(message, remoteInvokeError);
+});
+
+Then("writes console error {string}.", (error: string) => {
+    assert.equal(true, remoteInvokeConsoleError?.toString().includes(error));
+});
+
+Then("writes no console error.", () => {
+    assert.equal(undefined, remoteInvokeConsoleError);
+});
+
 Before(() => {
     rmSync(resolve(targetDirectory, ".."), { recursive: true, force: true });
     // eslint-disable-next-line no-console
     console.warn = () => undefined;
+
+    remoteInvokeConsoleError = undefined;
+    remoteInvokeError = undefined;
+    remoteInvokeResult = undefined;
 });
 
 After(() => {
     rmSync(resolve(targetDirectory, ".."), { recursive: true, force: true });
+    cli?.cleanup();
 });