+++ packages/test/test/__data__/faas-modules/echo.ts
@@ -1,0 +1,4 @@
+// @service()
+export const echo = async (echo: unknown) => {
+    return echo;
+};